requirements.txt
semantic-kernel>=1.41.0
azure-identity>=1.15.0
azure-monitor-opentelemetry>=1.2.0
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
python-dotenv>=1.0.0
jupyter>=1.0.0
ipykernel>=6.29.0
